[libc] Refactor thread.h
- extract ThreadAttributes and related types into a separate header.
  This is to allow it to be referenced from other places (like the
  thread control block structure I am planning to introduce) without
  pulling in the higher level dependencies and without circular
  includes.
- move `self` into the internal namespace and provide a
  `current_thread()` accessor, preparing the code to store the thread
  pointer inside the thread control block.
