+++ libc/src/__support/threads/CMakeLists.txt
@@ -19,6 +19,19 @@ add_header_library(
     libc.src.__support.CPP.atomic
 )
 
+add_header_library(
+  thread_attributes
+  HDRS
+    thread_attributes.h
+  DEPENDS
+    libc.hdr.stdint_proxy
+    libc.hdr.types.size_t
+    libc.src.__support.CPP.atomic
+    libc.src.__support.macros.attributes
+    libc.src.__support.macros.config
+    libc.src.__support.macros.properties.architectures
+)
+
 if(EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/${LIBC_TARGET_OS})
   add_subdirectory(${LIBC_TARGET_OS})
 endif()
@@ -111,6 +124,7 @@ add_header_library(
   HDRS
     thread.h
   DEPENDS
+    .thread_attributes
     libc.hdr.stdint_proxy
     libc.hdr.types.struct_sched_param
     libc.src.__support.common
