+++ libc/src/stdlib/exit.cpp
@@ -23,7 +23,7 @@ extern "C" void __cxa_finalize(void *);
 [[noreturn]] LLVM_LIBC_FUNCTION(void, exit, (int status)) {
 #ifdef LIBC_COPT_SUPPORT_THREADS
   // Call TLS destructors, if supported by the target.
-  internal::call_atexit_callbacks(self.attrib);
+  internal::call_atexit_callbacks(current_thread().attrib);
 #endif
   __cxa_finalize(nullptr);
   internal::exit(status);
